+++ doc/changes.rst
@@ -19,6 +19,12 @@ Minor changes
 =====================
 
 
+Minor changes
+-------------
+
+* Add an the alias `hide_drop_down` to DataValidation for `showDropDown` because that is how Excel works.
+
+
 Bug fixes
 ---------